src/requirements.txt
asgiref==3.3.1
pypi==2.1
gunicorn==20.0.4
psycopg2-binary==2.8.6
django-redis==4.12.1
Django==3.1.3
django-debug-toolbar==3.2
python-slugify==4.0.1
pytz==2020.4
sqlparse==0.4.1
text-unidecode==1.3
celery==5.0.5
